Add CurrencyApiModelExtensions file
